requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
langchain==0.0.350
langchain-openai>=0.0.2
langchain-core>=0.1.0
openai>=1.3.0
PyPDF2>=3.0.1
pandas>=2.1.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0

